% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subtype.R
\name{create_subtype}
\alias{create_subtype}
\title{Prepare a stratified phenotype for multivariate LDAK.}
\usage{
create_subtype(pheno = NULL, cov = NULL, size = 5)
}
\arguments{
\item{pheno}{Phenotype matrix or dataframe in PLINK format}

\item{cov}{Covariate vector in PLINK format}

\item{size}{Number of subgroups for stratifying the phenotype}
}
\value{
The output of the system call
}
\description{
The binary outcome is stratified based on quantiles of the
associated covariate.
}
