% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldak.R
\name{ldak}
\alias{ldak}
\title{Run LDAK executable for Linux, Mac, and Windows systems.}
\usage{
ldak(args = character())
}
\arguments{
\item{args}{Character vector of arguments to pass to LDAK}
}
\value{
The output of the system call
}
\description{
This function automatically detects the operating system and runs
the corresponding LDAK binary shipped with the package.
}
