% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_subtype.R
\name{make_subtype}
\alias{make_subtype}
\title{Prepare a stratified phenotype for multivariate LDAK.}
\usage{
make_subtype(pheno = NULL, cov = NULL, size = numeric())
}
\arguments{
\item{pheno}{Phenotype matrix or dataframe in PLINK format}

\item{cov}{Covariate vector}

\item{size}{Number of subgroups for stratifying the phenotype}
}
\value{
The output of the system call
}
\description{
The binary outcome is stratified based on quantiles of the
associated covariate.
}
