% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_transform.R
\name{make_transform}
\alias{make_transform}
\title{Compute transformed phenotype based on LDAK output.}
\usage{
make_transform(
  pheno = NULL,
  cov = NULL,
  size = numeric(),
  trans = NULL,
  spar = 0.2
)
}
\arguments{
\item{pheno}{Phenotype matrix or dataframe in PLINK format}

\item{cov}{Covariate vector}

\item{size}{Number of subgroups for stratifying the phenotype}

\item{trans}{Transformation matrix obtained from LDAK}

\item{spar}{Smoothing parameter (default: 0.2)}
}
\value{
A transformed phenotypes which can be used for GWAS analysis
}
\description{
The transformation are used in combination with the original
covariate to compute a smoothed phenotype for association analysis.
}
