% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldak.R
\name{run_ldak}
\alias{run_ldak}
\title{Run LDAK executable}
\usage{
run_ldak(args = character())
}
\arguments{
\item{args}{Character vector of arguments to pass to LDAK}
}
\value{
The output of the system call
}
\description{
This function automatically detects the operating system and runs
the corresponding LDAK binary shipped with the package.
}
